tests/el/android/assets/AssetsTest.java
package el.android.assets;

import android.app.Activity;
import android.content.res.AssetManager;
import android.graphics.Bitmap;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.Robolectric;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.annotation.Config;

import el.android.Game;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertNull;
import static org.junit.Assert.assertTrue;

/**
 * Created by ash on 16.06.16.
 */

//@Config(emulateSdk=18)
@Config(sdk = 18)
@RunWith(RobolectricTestRunner.class)
public class AssetsTest {

    AssetManager assetManager;

    @Before
    public void setUp() throws Exception {
        assetManager = Robolectric.buildActivity(Game.class).create().get().getAssets();
        assertNotNull(assetManager);
        Assets.setAssetManager(assetManager);
    }

    @Test
    public void testActivityFound() {
        Activity activity = Robolectric.buildActivity(Game.class).create().get();
        Assert.assertNotNull(activity);
    }

    @Test
    public void assertGetBitmapNotNull() {
        Bitmap bitmap = Assets.loadBitmap("items/items11.png");
        assertNotNull(bitmap);
    }

    @Test(expected = RuntimeException.class)
    public void assertGetBitmapNull() {
        Bitmap bitmap = Assets.loadBitmap("items/items0.png");
        assertNull(bitmap);
    }

    @Test
    public void assertBitmapSameAs() {
        Bitmap bitmap1 = Assets.loadBitmap("items/items1.png");
        Bitmap bitmap2 = Assets.loadBitmap("items/items1.png");
        // FIXME For now this does not work, not sure why
        assertFalse(bitmap1.sameAs(bitmap2));
    }

    @Test
    public void assertGetItemImageNotNull() {
        Assets.IconBitmap iconBitmap = Assets.getItemImage(1);
        System.out.println("bitmap 1 " + iconBitmap.bitmap + "x " + iconBitmap.x + "y " + iconBitmap.y);
        assertNotNull(iconBitmap);
    }

    @Test
    public void testIconBitmapSameAs() {

        Assets.IconBitmap iconBitmap1 = Assets.getItemImage(15);
        Assets.IconBitmap iconBitmap2 = Assets.getItemImage(15);

        assertTrue(iconBitmap1.sameAs(iconBitmap2));
    }

    @Test
    public void testIconBitmapNotSameAs() {
        // Test different bitmap
        Assets.IconBitmap iconBitmap1 = Assets.getItemImage(10);
        Assets.IconBitmap iconBitmap2 = Assets.getItemImage(35);

        assertFalse(iconBitmap1.sameAs(iconBitmap2));

        // Test different item location same bitmap
        Assets.IconBitmap iconBitmap3 = Assets.getItemImage(11);

        assertFalse(iconBitmap1.sameAs(iconBitmap3));

    }

    @Test
    public void assertGetItemImageDefault() {

        Assets.IconBitmap iconBitmap = Assets.getItemImage(1000);

        Assets.IconBitmap expected = Assets.getItemImage(27 * 25 - 1);

        assertTrue(iconBitmap.sameAs(expected));
    }

    @Test
    public void testGetItemImage() {
        Assets.IconBitmap iconBitmap = Assets.getItemImage(262);

        // Bitmap bitmap = Assets.loadBitmap("items/items11.png");
        Assets.IconBitmap expected = new Assets.IconBitmap(Assets.getItemBitmapFromFileId(10), 100, 100);

        assertTrue(iconBitmap.sameAs(expected));
    }
}
